Migrate the Resume database service  to the new app directory
